requirements.txt
pytest==7.4.0
pytest-asyncio==0.21.0
python-telegram-bot==20.7